api/auth/requirements.txt
aiohttp==3.6.2
aiohttp-jwt==0.5.0
aioredis==1.3.1
async-timeout==3.0.1
asyncpg==0.20.0
asyncpgsa==0.26.1
attrs==19.3.0
bcrypt==3.1.7
chardet==3.0.4
idna==2.8
multidict==4.7.1
psycopg2==2.8.4
PyJWT==1.7.1
PyYAML==5.2
SQLAlchemy==1.3.12
trafaret==2.0.2
yarl==1.4.2
